requirements.txt
Flask==2.2.3
Flask-WTF==1.1.1
WTForms==3.0.1
qrcode==7.4.2
Pillow==10.4.0
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Caching==2.0.2
Flask-Limiter==3.3.1
Flask-Cors==4.0.0
Flask-Dance==7.0.0
blinker==1.6.2
Werkzeug==2.2.3
python-dotenv==1.0.0
email_validator==2.0.0
geoip2==4.7.0
user-agents==2.2.0
psycopg2-binary==2.9.10
marshmallow==3.19.0
celery==5.3.1
redis==4.5.5
pytest==7.3.1
black==23.3.0
flake8==6.0.0
gunicorn==20.1.0
Faker==19.13.0
sendgrid
requests>=2.31.0
psutil>=5.9.0
twilio==9.5.1
